# refactor: increase performance by ~7% thanks to Tokens::block*Cache hit increased by ~12%