--- a/src/Tokenizer/Tokens.php
+++ b/src/Tokenizer/Tokens.php
@@ -314,23 +314,8 @@ class Tokens extends \SplFixedArray
      */
     public function offsetSet($index, $newval): void
     {
-        // # cache usage count with following variant: tests/Fixer/: 102_074
-        // $this->blockStartCache = [];
-        // $this->blockEndCache = [];
-
         if (!isset($this[$index]) || !$this[$index]->equals($newval)) {
-            // # cache usage count with following variant: tests/Fixer/: 108_464
-            // $this->blockStartCache = [];
-            // $this->blockEndCache = [];
-
-            if (isset($this[$index]) && self::isTokenBlockEdge($this[$index])) {
-                // # cache usage count with following variant: tests/Fixer/: 113_324
-                // $this->blockStartCache = [];
-                // $this->blockEndCache = [];
-            }
-
             if (isset($this[$index])) {
-                // # cache usage count with following variant: tests/Fixer/: 114_574
                 if (isset($this->blockStartCache[$index])) {
                     unset($this->blockEndCache[$this->blockStartCache[$index]], $this->blockStartCache[$index]);
                 }