### meta

- attempts: ~5 (most of them merged as a partial victory)
- mental execution time: ~2yrs (with partial merges over time, and also 2nd author)
- actual coding: ~1h (for this approach), 1h for cleanups around (extracted to other PRs)

### results

- approx ~7% performance gain
- increase Tokens::block*Cache hit by ~12%

### approach

- no need to clear the blockCache for such changes

```php
// $tokens[5] = [T_DOC_COMMENT, "/** Foo */"
$tokens[5] = new Token([T_COMMENT, "/* Foo */");
```

### benchmark

```
# ./php-cs-fixer fix -vvv # on master
Fixed 0 of 1063 files in 66.990 seconds, 24.000 MB memory used
Fixed 0 of 1063 files in 67.361 seconds, 24.000 MB memory used
Fixed 0 of 1063 files in 66.848 seconds, 24.000 MB memory used
```

```
# ./php-cs-fixer fix -vvv # on PR
Fixed 0 of 1063 files in 62.640 seconds, 24.000 MB memory used
Fixed 0 of 1063 files in 62.527 seconds, 24.000 MB memory used
Fixed 0 of 1063 files in 62.846 seconds, 24.000 MB memory used
```