@SpacePossum 2021-12-16
please see some findings in https://github.com/keradus/PHP-CS-Fixer/pull/34 :)